Move base64 helpers into @intx/types
The sidecar's WebSocket client carried a private pair of base64 codec
helpers used to ship binary mail bodies over text-only frames. They
are wire-format primitives shared with any peer on the same protocol,
not sidecar internals. Park them alongside the hex codec in
@intx/types and rename to base64Encode and base64Decode to match the
hexEncode and hexDecode shape.
